TCA-55 The start folder level was removed. And the activities were moved to presentation folder
