+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:supportsRtl="true"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.starter.LoginActivity"
+            android:name=".presentation.LoginActivity"
             android:exported="true"
             android:theme="@style/Theme.TwitterCloneApp">
             <intent-filter>
@@ -23,7 +23,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".presentation.main.MainActivity"
+            android:name=".presentation.MainActivity"
             android:theme="@style/Theme.TwitterCloneAppWithActionBar">
         </activity>
     </application>
