+++ app/src/main/java/ar/scacchipa/twittercloneapp/di/AppModule.kt
@@ -8,14 +8,14 @@ import ar.scacchipa.twittercloneapp.data.datasource.IAuthDataSource
 import ar.scacchipa.twittercloneapp.domain.usecase.AuthorizationUseCase
 import ar.scacchipa.twittercloneapp.domain.usecase.ConsumableAuthUseCase
 import ar.scacchipa.twittercloneapp.domain.usecase.SplashTimerUseCase
-import ar.scacchipa.twittercloneapp.presentation.starter.login.FragmentLoginWebSection
-import ar.scacchipa.twittercloneapp.presentation.starter.login.FragmentLogin
-import ar.scacchipa.twittercloneapp.presentation.starter.splash.FragmentSplash
+import ar.scacchipa.twittercloneapp.presentation.login.FragmentLoginWebSection
+import ar.scacchipa.twittercloneapp.presentation.login.FragmentLogin
+import ar.scacchipa.twittercloneapp.presentation.splash.FragmentSplash
 import ar.scacchipa.twittercloneapp.data.repository.AuthorizationRepository
 import ar.scacchipa.twittercloneapp.data.repository.IAuthorizationRepository
-import ar.scacchipa.twittercloneapp.presentation.starter.login.LoginWebSectionViewModel
-import ar.scacchipa.twittercloneapp.presentation.starter.login.LoginViewModel
-import ar.scacchipa.twittercloneapp.presentation.starter.splash.SplashViewModel
+import ar.scacchipa.twittercloneapp.presentation.login.LoginWebSectionViewModel
+import ar.scacchipa.twittercloneapp.presentation.login.LoginViewModel
+import ar.scacchipa.twittercloneapp.presentation.splash.SplashViewModel
 import org.koin.androidx.fragment.dsl.fragment
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
