+++ app/src/main/java/ar/scacchipa/twittercloneapp/presentation/LoginActivity.kt
@@ -1,4 +1,4 @@
-package ar.scacchipa.twittercloneapp.presentation.starter
+package ar.scacchipa.twittercloneapp.presentation
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
