+++ app/src/main/java/ar/scacchipa/twittercloneapp/presentation/login/LoginWebSectionViewModel.kt
@@ -1,4 +1,4 @@
-package ar.scacchipa.twittercloneapp.presentation.starter.login
+package ar.scacchipa.twittercloneapp.presentation.login
 
 import android.webkit.WebResourceError
 import android.webkit.WebViewClient
