+++ app/src/main/res/layout/activity_login_layout.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.starter.LoginActivity">
+    tools:context=".presentation.LoginActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/login_host_fragment"
