+++ app/src/main/res/layout/fragment_login_web_section_layout.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/secondary_background"
-    tools:context=".presentation.starter.login.FragmentLoginWebSection">
+    tools:context="ar.scacchipa.twittercloneapp.presentation.login.FragmentLoginWebSection">
 
     <WebView
         android:id="@+id/login_webview"
