+++ app/src/main/res/layout/fragment_splash_layout.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/primary_background"
-    tools:context=".presentation.starter.splash.FragmentSplash">
+    tools:context=".presentation.splash.FragmentSplash">
 
     <ImageView
         android:id="@+id/twitter_logo_imageview"
